Exempt query cancellation in MergeTreeSequentialSource's broken-part catch
MergeTreeSequentialSource::generate reports a suspected broken part for every
non-retryable exception, and isRetryableException does not recognise the three
cancellation codes. Once the compact reader's mark loop can throw them, a
cancelled foreground query enqueues a part check for a healthy Compact part.

MergeTreeDataPartCompact::getReader returns MergeTreeReaderCompactSingleBuffer,
so a Compact read through this source enters the loop this PR changed. The
foreground caller is WhatIfEmpiricalEstimator, which drives the source from
EXPLAIN WHATIF on the calling thread with the query's process-list element
attached and no experimental setting gating it; the other three callers are
merges and mutations. reportBroken reaches broken_part_callback, which for
ReplicatedMergeTree queues the part for a full data check.

Measured on a Replicated Compact table with index_granularity = 1 and 400k
marks. With both sites instrumented, a deadline landing inside the read
reached the catch twice and ran reportBroken both times, and the server then
logged a full re-check of the active part ending "Part all_0_0_0 looks good."
Reverting only the new conjunct restores that: on a paired ladder where each
arm reached the read once, the pre-fix arm enqueued one check and the fixed arm
enqueued none, with the part active and not detached in both.

The classifier moves from the compact reader's anonymous namespace into
checkDataPart, beside isRetryableException, and both catch sites call it, so it
is defined once rather than twice. Both files already included that header for
isRetryableException, so this adds no dependency. Adding the codes to
isRetryableException itself would have been wrong: its three call sites inside
checkDataPart treat retryable as "return empty checksums and retry later".

A loop-top deadline check in this source was measured dead in an earlier round
and stays out: readRows is called with one mark's rows, so its loop body runs
at most once. That is also why the defect is narrow, since a deadline usually
fires between marks. Reachability of a regression arm for the wiring measured
1 in 20, so no arm is added rather than a flaky one; the classifier itself is
covered by the existing arms of 04746.
