+++ src/Storages/MergeTree/checkDataPart.h
@@ -17,4 +17,8 @@ IMergeTreeDataPart::Checksums checkDataPart(
 bool isNotEnoughMemoryErrorCode(int code);
 bool isRetryableException(std::exception_ptr exception_ptr);
 
+/// A cancelled or timed-out query says nothing about the part's health, so such an exception must
+/// not reach reportBroken.
+bool isQueryCancellation(std::exception_ptr exception_ptr);
+
 }
